Phase 2 Complete Work
Successfully updated the app to store the retrieved facts in a SQLite database using Room. I utilized ChatGPT with a couple of weird errors such as the JSON not playing nice with id in the new Entity. Other than that I was able to piece through the Weather Example Code and get a very good understanding for how this setup works.